+++ Eshava.DomainDrivenDesign.CodeAnalysis/Models/Application/UseCaseTemplateRequest.cs
@@ -16,6 +16,7 @@ namespace Eshava.DomainDrivenDesign.CodeAnalysis.Models.Application
 		public UseCasesMap UseCasesMap { get; set; }
 		public List<UseCaseCodeSnippet> CodeSnippets { get; set; }
 		public List<ApplicationProjectAlternativeClass> AlternativeClasses { get; set; }
+		public bool UseNewtonsoftJson { get; set; }
 		public bool AddAssemblyCommentToFiles { get; set; }
 	}
 }